Split out action driver code
